+++ History.md
@@ -1,5 +1,11 @@
 History
 ===
 
+##### 2016-06-21 : v0.2.0
+Try to figure out whether res is http2/spdy response,
+if so, remove legacy HTTP/1.x headers:
+
+`connection, keep-alive, proxy-connection, transfer-encoding, upgrade`
+
 ##### 2016-05-10 : v0.1.0
 Initial commit